backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
aiomysql==0.2.0
cryptography==41.0.7
email-validator==2.1.0
fastapi-csrf-protect==2.1.0  # NOVO
resend
pyotp
redis 
aioredis 
slowapi